+++ src/Channels/ThreemaChannel.php
@@ -3,7 +3,13 @@
 namespace Illuminate\Notifications\Channels;
 
 use Illuminate\Notifications\Exceptions\ThreemaChannelException;
+use Illuminate\Notifications\Exceptions\ThreemaChannelResultException;
+use Illuminate\Notifications\Messages\ThreemaAudioMessage;
+use Illuminate\Notifications\Messages\ThreemaFileMessage;
+use Illuminate\Notifications\Messages\ThreemaImageMessage;
+use Illuminate\Notifications\Messages\ThreemaLocationMessage;
 use Illuminate\Notifications\Messages\ThreemaTextMessage;
+use Illuminate\Notifications\Messages\ThreemaVideoMessage;
 use Illuminate\Notifications\Notification;
 use Threema\MsgApi\Commands\Results\Result;
 use Threema\MsgApi\Connection;
@@ -37,19 +43,20 @@ class ThreemaChannel
 
 	/**
 	 * @throws ThreemaChannelException
+	 * @throws ThreemaChannelResultException
 	 */
-	public function send(mixed $notifiable, Notification $notification): ?Result
+	public function send(mixed $notifiable, Notification $notification): Result
 	{
 		$message = $notification->toThreema($notifiable);
-		$connection = $message->channel ? $message->channel->connection : $this->connection;
-		$privateKey = $message->channel ? $message->channel->privateKey : $this->privateKey;
+		$connection = $message->getChannel() ? $message->getChannel()->connection : $this->connection;
+		$privateKey = $message->getChannel() ? $message->getChannel()->privateKey : $this->privateKey;
 
 		if (!$receiver = $notifiable->routeNotificationFor('threema', $notification)) {
 			throw new ThreemaChannelException('Notifiable is missing "routeNotificationForThreema" function.');
 		}
 
 		if ($privateKey === null && $message instanceof ThreemaTextMessage) {
-			$result = $connection->sendSimple($receiver, $message->text);
+			$result = $connection->sendSimple($receiver, $message->getText());
 		} else {
 			$e2eHelper = new E2EHelper($privateKey, $connection);
 
@@ -67,7 +74,7 @@ class ThreemaChannel
 					if ($lookup->isSuccess()) {
 						$threemaId = $lookup->getId();
 					} else {
-						return $lookup;
+						throw new ThreemaChannelResultException($lookup);
 					}
 				} else {
 					throw new ThreemaChannelException('This lookup type is not supported by Laravel Threema.');
@@ -76,13 +83,46 @@ class ThreemaChannel
 
 			try {
 				if ($message instanceof ThreemaTextMessage) {
-					$result = $e2eHelper->sendTextMessage($threemaId, $message->text);
+					$result = $e2eHelper->sendTextMessage($threemaId, $message->getText());
+				} else if ($message instanceof ThreemaImageMessage) {
+					$result = $e2eHelper->sendImageMessage($threemaId, $message->getPath());
+				} else if ($message instanceof ThreemaLocationMessage) {
+					$result = $e2eHelper->sendLocationMessage(
+						$threemaId,
+						$message->getLatitude(),
+						$message->getLongitude(),
+						$message->getAccuracy(),
+						$message->getPoiName(),
+						$message->getPoiAddress()
+					);
+				} else if ($message instanceof ThreemaAudioMessage) {
+					$result = $e2eHelper->sendAudioMessage($threemaId, $message->getDuration(), $message->getPath());
+				} else if ($message instanceof ThreemaVideoMessage) {
+					$result = $e2eHelper->sendVideoMessage(
+						$threemaId,
+						$message->getDuration(),
+						$message->getPath(),
+						$message->getThumbnailPath(),
+					);
+				} else if ($message instanceof ThreemaFileMessage) {
+					$result = $e2eHelper->sendFileMessage(
+						$threemaId,
+						$message->getPath(),
+						$message->getThumbnailPath(),
+						$message->getCaption(),
+						$message->getName(),
+						$message->getType(),
+					);
 				} else {
 					throw new ThreemaChannelException('This message type is not supported by Laravel Threema.');
 				}
 			} catch (Exception $exception) {
 				throw new ThreemaChannelException('The underlying Threema MsgApi has thrown an exception.', 0, $exception);
 			}
+		}
+
+		if(!$result->isSuccess()) {
+			throw new ThreemaChannelResultException($result);
 		}
 
 		return $result;