+++ src/Messages/ThreemaMessage.php
@@ -6,10 +6,22 @@ use Illuminate\Notifications\Channels\ThreemaChannel;
 
 abstract class ThreemaMessage
 {
-	public ?ThreemaChannel $channel;
+	protected ?ThreemaChannel $channel;
 
 	public function __construct(?ThreemaChannel $channel = null)
 	{
 		$this->channel = $channel;
+	}
+
+	public function getChannel(): ?ThreemaChannel
+	{
+		return $this->channel;
+	}
+
+	public function setChannel(?ThreemaChannel $channel): self
+	{
+		$this->channel = $channel;
+
+		return $this;
 	}
 }