+++ src/Messages/ThreemaTextMessage.php
@@ -6,12 +6,24 @@ use Illuminate\Notifications\Channels\ThreemaChannel;
 
 class ThreemaTextMessage extends ThreemaMessage
 {
-	public string $text;
+	protected string $text;
 
 	public function __construct(string $text, ?ThreemaChannel $channel = null)
 	{
 		parent::__construct($channel);
 
 		$this->text = $text;
+	}
+
+	public function getText(): string
+	{
+		return $this->text;
+	}
+
+	public function setText(string $text): self
+	{
+		$this->text = $text;
+
+		return $this;
 	}
 }